// src/app/shared/models/user.model.ts
export class UserModel{
    id: string;
    first_name: string;
    last_name: string;
    email: string;
    contact_number: string;
    address1: string;
    address2: string;
    city: string;
    province: string;
    country: string;
    postal_code: string;
    password: string;
    access_level: string;
    access_code: string;
    status: number;
    created: Date;
    modified: Date;
}